backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.121.0
uvicorn==0.38.0
starlette==0.49.3
h11==0.16.0

# Database and ORM
SQLAlchemy==2.0.44
alembic==1.15.0
greenlet==3.2.4

# Authentication and Security
passlib==1.7.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.18

# Data Validation and Serialization
pydantic==2.12.3
pydantic_core==2.41.4
email-validator==2.3.0

# Utilities
python-dotenv==1.0.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
idna==3.11
pycparser==2.23
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0

# Optional: For production deployment
gunicorn==23.0.0

# Optional: For API documentation enhancement
markdown==3.7

# Optional: For CORS and additional middleware
python-jose[cryptography]==3.3.0

# LangChain and AI/LLM Integration
langchain==0.3.13
langchain-google-genai==2.0.8
google-generativeai==0.8.3
langchain-community==0.3.13
chromadb==0.5.23
